Add files via upload
TEV is a little doohickey I made to validate transpeople and it is technically my python homework since it uses all of the python I know in one program. This version is only 66 lines, I plan to continue to expand it as I learn more. I will publish more versions as time wears on. Open this in the command prompt and answer with single words. If it gives you specific options you have to pick one. If it's a yes or no question, you have to type yes or no. Feel free to explore the program. Please playthrough at least once before you look at the source code.